Rename gate package to reflect component
