+++ frontend/src/app/gate/gate.component.ts
